Added Databrowser
Iterates through pickle files and displays data, description, comment
and image (if these exist) of each file in dataframe table. For the
images to work, I changed the save function in the plotting thread to
create a renderer which saves as a png. However according to
(https://github.com/ioam/holoviews/issues/2151), this uses selenium and
is relatively slow (takes a couple seconds) so might want to make this
optional. In general this might be fine because saving is done in
background thread, but careful of issues due to this.