Further instrument agnostic changes
Now with measurement instrument that has measure method, and instruments
with ramp methods, can add each to overview class and do sweeps of
different instruments. Current things to fix are saving the state of the
system (right now tries to get metadata of qdac which is nonexistent).
With QCodes can take advantage of snapshot function